rest improvements, extjs5 improvments

+++ src/Phlexible/Bundle/SiterootBundle/Controller/SiterootsController.php
@@ -8,19 +8,16 @@
 
 namespace Phlexible\Bundle\SiterootBundle\Controller;
 
-use FOS\RestBundle\Controller\Annotations\NamePrefix;
-use FOS\RestBundle\Controller\Annotations\Prefix;
-use FOS\RestBundle\Controller\Annotations\Put;
-use FOS\RestBundle\Controller\Annotations\View;
+use FOS\RestBundle\Controller\Annotations as Rest;
 use FOS\RestBundle\Controller\FOSRestController;
+use FOS\RestBundle\View\View;
 use Nelmio\ApiDocBundle\Annotation\ApiDoc;
 use Phlexible\Bundle\SiterootBundle\Entity\Siteroot;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
-use Sensio\Bundle\FrameworkExtraBundle\Configuration\Route;
+use Phlexible\Bundle\SiterootBundle\Form\Type\SiterootType;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
-use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 /**
  * Siteroot controller
@@ -28,8 +25,7 @@ use Symfony\Component\HttpFoundation\Response;
  * @author Stephan Wentz <sw@brainbits.net>
  *
  * @Security("is_granted('ROLE_SITEROOTS')")
- * @Prefix("/siteroot")
- * @NamePrefix("phlexible_siteroot_")
+ * @Rest\NamePrefix("phlexible_api_siteroot_")
  */
 class SiterootsController extends FOSRestController
 {
@@ -38,7 +34,15 @@ class SiterootsController extends FOSRestController
      *
      * @return Response
      *
-     * @ApiDoc
+     * @Rest\View
+     * @ApiDoc(
+     *   description="Returns a collection of Siteroot",
+     *   section="siteroot",
+     *   resource=true,
+     *   statusCodes={
+     *     200="Returned when successful",
+     *   }
+     * )
      */
     public function getSiterootsAction()
     {
@@ -46,12 +50,10 @@ class SiterootsController extends FOSRestController
 
         $siteroots = $siterootManager->findAll();
 
-        return $this->handleView($this->view(
-            array(
-                'siteroots' => $siteroots,
-                'count' => count($siteroots)
-            )
-        ));
+        return array(
+            'siteroots' => $siteroots,
+            'count'     => count($siteroots)
+        );
     }
 
     /**
@@ -61,63 +63,119 @@ class SiterootsController extends FOSRestController
      *
      * @return Response
      *
-     * @View(templateVar="siteroot")
-     * @ApiDoc
+     * @Rest\View
+     * @ApiDoc(
+     *   description="Returns a Siteroot",
+     *   section="siteroot",
+     *   output="Phlexible\Bundle\SiterootBundle\Entity\Siteroot",
+     *   statusCodes={
+     *     200="Returned when successful",
+     *     404="Returned when siteroot was not found"
+     *   }
+     * )
      */
     public function getSiterootAction($siterootId)
     {
         $siterootManager = $this->get('phlexible_siteroot.siteroot_manager');
-
         $siteroot = $siterootManager->find($siterootId);
 
-        return $siteroot;
+        if (!$siteroot instanceof Siteroot) {
+            throw new NotFoundHttpException('Siteroot not found');
+        }
+
+        return array(
+            'siteroot' => $siteroot
+        );
     }
 
     /**
      * Create siteroot
      *
-     * @param Siteroot $siteroot
+     * @param Request $request
      *
      * @return Response
      *
-     * @ParamConverter("siteroot", converter="fos_rest.request_body")
-     * @ApiDoc
+     * @ApiDoc(
+     *   description="Create a Siteroot",
+     *   section="siteroot",
+     *   input="Phlexible\Bundle\SiterootBundle\Form\Type\SiterootType",
+     *   statusCodes={
+     *     201="Returned when siteroot was created",
+     *     204="Returned when siteroot was updated",
+     *     404="Returned when siteroot was not found"
+     *   }
+     * )
      */
-    public function postSiterootsAction(Siteroot $siteroot)
+    public function postSiterootsAction(Request $request)
     {
-        $siterootManager = $this->get('phlexible_siteroot.siteroot_manager');
-
-        $siterootManager->updateSiteroot($siteroot);
-
-        return $this->handleView($this->view(
-            array(
-                'success' => true,
-            )
-        ));
+        return $this->processForm($request, new Siteroot());
     }
 
     /**
      * Update siteroot
      *
-     * @param Siteroot $siteroot
+     * @param Request $request
+     * @param string  $siterootId
      *
      * @return Response
      *
-     * @ParamConverter("siteroot", converter="fos_rest.request_body")
-     * @Put("/siteroots/{siterootId}")
-     * @ApiDoc
+     * @ApiDoc(
+     *   description="Update a Siteroot",
+     *   section="siteroot",
+     *   input="Phlexible\Bundle\SiterootBundle\Form\Type\SiterootType",
+     *   statusCodes={
+     *     201="Returned when siteroot was created",
+     *     204="Returned when siteroot was updated",
+     *     404="Returned when siteroot was not found"
+     *   }
+     * )
      */
-    public function putSiterootAction(Siteroot $siteroot, $siterootId)
+    public function putSiterootAction(Request $request, $siterootId)
     {
         $siterootManager = $this->get('phlexible_siteroot.siteroot_manager');
+        $siteroot = $siterootManager->find($siterootId);
 
-        $siterootManager->updateSiteroot($siteroot);
+        if (!$siteroot instanceof Siteroot) {
+            throw new NotFoundHttpException('Siteroot not found');
+        }
 
-        return $this->handleView($this->view(
-            array(
-                'success' => true,
-            )
-        ));
+        return $this->processForm($request, $siteroot);
+    }
+
+    /**
+     * @param Request  $request
+     * @param Siteroot $siteroot
+     *
+     * @return Rest\View|Response
+     */
+    private function processForm(Request $request, Siteroot $siteroot)
+    {
+        $statusCode = !$siteroot->getId() ? 201 : 204;
+
+        $form = $this->createForm(new SiterootType(), $siteroot);
+        $form->submit($request);
+
+        if ($form->isValid()) {
+            $siterootManager = $this->get('phlexible_siteroot.siteroot_manager');
+            $siterootManager->updateSiteroot($siteroot);
+
+            $response = new Response();
+            $response->setStatusCode($statusCode);
+
+            // set the `Location` header only when creating new resources
+            if (201 === $statusCode) {
+                $response->headers->set('Location',
+                    $this->generateUrl(
+                        'phlexible_api_siteroot_get_siteroot', array('siterootId' => $siteroot->getId()),
+                        true // absolute
+                    )
+                );
+            }
+
+            return $response;
+        }
+
+        return View::create($form, 400);
     }
 
     /**
@@ -127,19 +185,25 @@ class SiterootsController extends FOSRestController
      *
      * @return Response
      *
-     * @ApiDoc
+     * @Rest\View(statusCode=204)
+     * @ApiDoc(
+     *   description="Delete a Siteroot",
+     *   section="siteroot",
+     *   statusCodes={
+     *     204="Returned when successful",
+     *     404="Returned when the siteroot is not found"
+     *   }
+     * )
      */
     public function deleteSiterootAction($siterootId)
     {
         $siterootManager = $this->get('phlexible_siteroot.siteroot_manager');
-
         $siteroot = $siterootManager->find($siterootId);
-        $siterootManager->deleteSiteroot($siteroot);
 
-        return $this->handleView($this->view(
-            array(
-                'success' => true,
-            )
-        ));
+        if (!$siteroot instanceof Siteroot) {
+            throw new NotFoundHttpException('Siteroot not found');
+        }
+
+        $siterootManager->deleteSiteroot($siteroot);
     }
 }
